+++ test/e2e/_utils.ts
@@ -1,7 +1,7 @@
 import * as express from 'express';
 import { Express, RequestHandler } from 'express';
 
-export { createProxyMiddleware } from '../../dist/index';
+export { createProxyMiddleware, responseInterceptor } from '../../dist/index';
 
 export function createApp(middleware: RequestHandler): Express {
   const app = express();